# [DeepSeek][Kernels] add on device, parallelized permute_indices kernel