--- a/torchtitan/experiments/deepseek_v3/moe_kernels.py
+++ b/torchtitan/experiments/deepseek_v3/moe_kernels.py
@@ -4,8 +4,6 @@
 # This source code is licensed under the BSD-style license found in the
 # LICENSE file in the root directory of this source tree.
 
-from typing import Tuple
-
 import torch
 import triton
 import triton.language as tl