--- a/torchtitan/experiments/deepseek_v3/moe_kernels.py
+++ b/torchtitan/experiments/deepseek_v3/moe_kernels.py
@@ -146,8 +146,6 @@ def generate_permute_indices(
     max_len: int,
     alignment: int,
     use_cpu: bool = False,
-    block_size: int = 128,
-    max_blocks: int = 1024,
 ):
     """
     Prepare permutation indices and the number of tokens for each expert.
@@ -201,7 +199,6 @@ def generate_permute_indices(
             experts_per_rank,
             num_ranks,
             max_len,
-            block_size=block_size,
         )
 
     return permuted_indices, m_sizes