This PR updates the original Triton kernel in indices.py (https://github.com/pytorch/torchtitan/pull/1062) to move it to a more parallelized and vectorized implementation to better support larger runs.  
This provides roughly an 8.7x perf improvement for 256 experts. 

This kernel functionally achieves the same as before:
Prepare permutation indices and the number of tokens for each expert with any required alignment, where alignment is driven by the group gemm requirement. 
* The permutation indices are the indices of the tokens for each expert. 
* The number of tokens for each expert is the sum of the number of tokens for
such experts from all ranks. 

However, this kernel improves the parallelization by assigning one block to each expert, and threads in each block process the tokens in chunks, all while using vectorized loads and stores.   
This results in the performance improvement while yielding the same indices.

Testing:
~~~
python moe_kernels.py
  ~~~

(runs simple verification using cpu, previous kernel and new optimized kernel)

Benchmark of performance improvement using fixed 256 experts across various world sizes
~~~

Summary for fixed total experts (256), where we mimic various DeepSeek experts across a number of ranks:
Ranks | Experts/Rank | Original (ms) | Optimized (ms) | Block Size | Speedup | Correct
------------------------------------------------------------------------------------------
    1 |          256 |        0.241 |        0.020 |        128 |   11.95x |       1
    4 |           64 |        0.222 |        0.020 |        128 |   10.97x |       1
   16 |           16 |        0.221 |        0.021 |        128 |   10.40x |       1
   32 |            8 |        0.224 |        0.025 |         64 |    8.81x |       1
   64 |            4 |        0.225 |        0.037 |        128 |    6.01x |       1
  128 |            2 |        0.217 |        0.051 |        128 |    4.26x |       1
  ~~~

added unit testing: 
~~~
..Testing 256 experts per rank across 1 ranks (total: 256)
Testing 128 experts per rank across 2 ranks (total: 256)
Testing 64 experts per rank across 4 ranks (total: 256)
Testing 32 experts per rank across 8 ranks (total: 256)
Testing 16 experts per rank across 16 ranks (total: 256)
Testing 8 experts per rank across 32 ranks (total: 256)
Testing 4 experts per rank across 64 ranks (total: 256)
Testing 2 experts per rank across 128 ranks (total: 256)
.
----------------------------------------------------------------------
Ran 3 tests in 1.104s

OK
~~~